+++ frontend/src/components/routerpages/FAQ.js
